pyproject.toml
[project]
name = "corp-rfp-agent"
version = "0.3.0"
description = "AI-powered RFP answering system for Blue Yonder solutions"
requires-python = ">=3.10"

dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
